+++ src/dependency-builder/detectives/parser-helper.ts
@@ -26,6 +26,16 @@ export function getDependenciesFromCallExpression(node) {
   return null;
 }
 
+export function getSpecifierValueForImportDeclaration(specifier) {
+  return {
+    isDefault: specifier.type === 'ImportDefaultSpecifier',
+    // syntax of `import x from 'file'` doesn't have specifier.imported, only specifier.local
+    // syntax of `import { x as y } from 'file'`, has `x` as specifier.imported and `y` as
+    // specifier.local. we interested in `x` in this case.
+    name: specifier.imported ? specifier.imported.name : specifier.local.name
+  };
+}
+
 function getStringValue(node) {
   // using single or double quotes (', ")
   if (node.type === 'Literal' || node.type === 'StringLiteral') {