+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [unreleased]
+
+## [2.1.4-dev.1] - 2019-11-18
+
+- [#2140](https://github.com/teambit/bit/issues/2140) support `import { x as y }` syntax
+
 ## [2.1.3] - 2019-10-23
 
 ### Bug fixes